eliminacion de codigo duplicado en api ccn
